Add firebase staging configuration (#330)
