+++ client/package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "scripts": {
     "start": "cross-env VITE_BASE_API_URL=http://localhost:3000 vite",
-    "build": "vite build",
+    "build": "vite build --mode production",
+    "build-staging": "vite build --mode staging",
     "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "generate:api:client": "npx @openapitools/openapi-generator-cli generate -i ../openapi.yaml -g typescript-fetch -o src/api-client/ --skip-validate-spec",
     "preview": "vite preview",
