+++ client/src/init-firebase.ts
@@ -2,17 +2,10 @@ import { initializeApp } from 'firebase/app';
 import { getMessaging, getToken, onMessage } from 'firebase/messaging';
 import { connectAuthEmulator, getAuth } from 'firebase/auth';
 import { api } from './Config';
+import firebaseConfig from './firebase-config';
 
 export const initFirebaseApp = () => {
-  initializeApp({
-    apiKey: 'AIzaSyAo5AZ1Na4l0YlfhZAlIr0FaLH_S4_1gfM',
-    authDomain: 'ezla-pickup.firebaseapp.com',
-    projectId: 'ezla-pickup',
-    storageBucket: 'ezla-pickup.appspot.com',
-    messagingSenderId: '708652536157',
-    appId: '1:708652536157:web:53f1739dee1c5453eb58b4',
-    measurementId: 'G-EK4J3MGLJ8'
-  });
+  initializeApp(firebaseConfig[`${import.meta.env.MODE}`].options);
 
   if (import.meta.env.DEV) {
     const auth = getAuth();
@@ -53,8 +46,7 @@ export const setNotificationsToken = async () => {
       const messaging = getMessaging();
       // console.log(messaging);
       const fcmToken = await getToken(messaging, {
-        vapidKey:
-          'BCuU7yCkK_syBZI-Wm-aaUZJsnQkcLYlNBKxFzfvsEO03fX-JPJRPFLX_mufz9oqx0Q_cIWkbQc1IFQ_iRmcLl4'
+        vapidKey: firebaseConfig[`${import.meta.env.MODE}`].vapidKey
       });
 
       // Set token in our local storage and send to our server
